Refactor command logic into worker classes
