+++ src/main/java/org/javacord/bot/commands/migrated/MigratedToSlashCommandsNotifier.java
@@ -13,9 +13,8 @@ import org.javacord.api.interaction.SlashCommand;
 import org.javacord.bot.Constants;
 import org.javacord.bot.listeners.CommandCleanupListener;
 import org.javacord.bot.restrictions.Before2023;
+import org.javacord.bot.util.JavacordIconProvider;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.Comparator;
 
 @ApplicationScoped
@@ -35,37 +34,31 @@ import java.util.Comparator;
 //TODO: After 2022 delete this class, command cleanup listener and restriction classes
 public class MigratedToSlashCommandsNotifier implements Command<Message> {
     @Inject
-    Logger logger;
+    JavacordIconProvider iconProvider;
 
     @Override
     public void execute(CommandContext<? extends Message> commandContext) {
         Message message = commandContext.getMessage();
-        try (InputStream javacord3Icon = getClass().getResourceAsStream("/javacord3_icon.png")) {
-            EmbedBuilder embed = new EmbedBuilder()
-                    .setThumbnail(javacord3Icon)
-                    .setColor(Constants.JAVACORD_ORANGE)
-                    .setTitle("Migrated to Slash Commands")
-                    .setDescription("This bot was migrated to slash commands, just type `/` to discover them");
 
-            message
-                    .getApi()
-                    .getGlobalSlashCommands()
-                    .thenAccept(slashCommands -> slashCommands
-                            .stream()
-                            .sorted(Comparator.comparing(SlashCommand::getName))
-                            .forEachOrdered(slashCommand -> embed.addField(
-                                    String.format("</%s:%d>", slashCommand.getName(), slashCommand.getId()),
-                                    slashCommand.getDescription())
-                            ))
-                    .join();
+        EmbedBuilder embed = new EmbedBuilder()
+                .setThumbnail(iconProvider.getIcon())
+                .setColor(Constants.JAVACORD_ORANGE)
+                .setTitle("Migrated to Slash Commands")
+                .setDescription("This bot was migrated to slash commands, just type `/` to discover them");
 
-            CommandCleanupListener.insertResponseTracker(embed, message.getId());
-            message.reply(embed).join();
-        } catch (IOException ioe) {
-            logger
-                    .atError()
-                    .withThrowable(ioe)
-                    .log("Exception while closing image stream");
-        }
+        message
+                .getApi()
+                .getGlobalSlashCommands()
+                .thenAccept(slashCommands -> slashCommands
+                        .stream()
+                        .sorted(Comparator.comparing(SlashCommand::getName))
+                        .forEachOrdered(slashCommand -> embed.addField(
+                                String.format("</%s:%d>", slashCommand.getName(), slashCommand.getId()),
+                                slashCommand.getDescription())
+                        ))
+                .join();
+
+        CommandCleanupListener.insertResponseTracker(embed, message.getId());
+        message.reply(embed).join();
     }
 }
